refactor: move renderGrimoireHistory function to history module and update its usage
