+++ script.js
@@ -5,7 +5,7 @@ import { createCurvedLabelSvg, createDeathRibbonSvg } from './ui/svg.js';
 import { initSidebarResize, initSidebarToggle } from './ui/sidebar.js';
 import { initInAppTour } from './ui/tour.js';
 import { repositionPlayers as repositionPlayersLayout, positionRadialStack as positionRadialStackLayout } from './ui/layout.js';
-import { renderScriptHistory, saveHistories, loadHistories } from './ui/history.js';
+import { renderScriptHistory, saveHistories, loadHistories, renderGrimoireHistory } from './ui/history.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   const startGameBtn = document.getElementById('start-game');
@@ -124,45 +124,6 @@ document.addEventListener('DOMContentLoaded', () => {
     grimoireHistory: []
   };
   let isRestoringState = false;
-
-  function renderGrimoireHistory() {
-    if (!grimoireHistoryList) return;
-    grimoireHistoryList.innerHTML = '';
-    history.grimoireHistory.forEach(entry => {
-      const li = document.createElement('li');
-      li.dataset.id = entry.id;
-      li.className = 'history-item';
-      const nameSpan = document.createElement('span');
-      nameSpan.className = 'history-name';
-      nameSpan.textContent = entry.name || formatDateName(new Date(entry.createdAt || Date.now()));
-      // Inline edit input (hidden by default)
-      const nameInput = document.createElement('input');
-      nameInput.type = 'text';
-      nameInput.className = 'history-edit-input';
-      nameInput.value = entry.name || formatDateName(new Date(entry.createdAt || Date.now()));
-      nameInput.style.display = 'none';
-      // Icons
-      const renameBtn = document.createElement('button');
-      renameBtn.className = 'icon-btn rename';
-      renameBtn.title = 'Rename';
-      renameBtn.innerHTML = '<i class="fa-solid fa-pen"></i>';
-      const saveBtn = document.createElement('button');
-      saveBtn.className = 'icon-btn save';
-      saveBtn.title = 'Save';
-      saveBtn.style.display = 'none';
-      saveBtn.innerHTML = '<i class="fa-solid fa-check"></i>';
-      const deleteBtn = document.createElement('button');
-      deleteBtn.className = 'icon-btn delete';
-      deleteBtn.title = 'Delete';
-      deleteBtn.innerHTML = '<i class="fa-solid fa-trash"></i>';
-      li.appendChild(nameSpan);
-      li.appendChild(nameInput);
-      li.appendChild(renameBtn);
-      li.appendChild(saveBtn);
-      li.appendChild(deleteBtn);
-      grimoireHistoryList.appendChild(li);
-    });
-  }
 
   function addScriptToHistory(name, data) {
     const entryName = (name && String(name).trim()) || 'Custom Script';
@@ -534,7 +495,7 @@ document.addEventListener('DOMContentLoaded', () => {
       };
       history.grimoireHistory.unshift(entry);
       saveHistories(history);
-      renderGrimoireHistory();
+      renderGrimoireHistory({ grimoireHistoryList, history });
     } catch (_) { }
   }
 
@@ -669,7 +630,7 @@ document.addEventListener('DOMContentLoaded', () => {
         if (confirm('Delete this grimoire snapshot?')) {
           history.grimoireHistory = history.grimoireHistory.filter(x => x.id !== id);
           saveHistories(history);
-          renderGrimoireHistory();
+          renderGrimoireHistory({ grimoireHistoryList, history });
         }
         return;
       }
@@ -694,7 +655,7 @@ document.addEventListener('DOMContentLoaded', () => {
           entry.name = newName;
           entry.updatedAt = Date.now();
           saveHistories(history);
-          renderGrimoireHistory();
+          renderGrimoireHistory({ grimoireHistoryList, history });
         }
         li.classList.remove('editing');
         return;
@@ -734,7 +695,7 @@ document.addEventListener('DOMContentLoaded', () => {
           entry.name = newName;
           entry.updatedAt = Date.now();
           saveHistories(history);
-          renderGrimoireHistory();
+          renderGrimoireHistory({ grimoireHistoryList, history });
         }
       }
     });
@@ -2062,7 +2023,7 @@ document.addEventListener('DOMContentLoaded', () => {
   // Load histories and render lists
   loadHistories(history);
   renderScriptHistory({ scriptHistoryList, history });
-  renderGrimoireHistory();
+  renderGrimoireHistory({ grimoireHistoryList, history });
 
   // Restore previous session (script and grimoire)
   loadAppState();
